have a working image plot
Notes: The spherical harmonics in SciPy have theta and phi interchanged relative to how physicists think about them! Have a working image plot; not need to generalize code to create specify the image plane more generally; would be nice to have a 3d surface of constant prob (with a user-specified value) AND a slice through the surface.